feat: 충전 한도 타입 수정 - 이후 확장성을 위해 int에서 long으로 변경

+++ src/main/java/org/c4marathon/assignment/dto/request/PostMainAccountReq.java
@@ -7,6 +7,6 @@ public record PostMainAccountReq(
 	long userId,
 
 	@Positive(message = "이체 금액은 양수가 되어야 합니다.")
-	int amount
+	long amount
 ) {
 }
